+++ eno-core/src/main/java/fr/insee/eno/core/processing/common/steps/EnoAddPrefixInQuestionLabels.java
@@ -12,7 +12,7 @@ import fr.insee.eno.core.parameter.EnoParameters;
 import fr.insee.eno.core.parameter.EnoParameters.QuestionNumberingMode;
 import fr.insee.eno.core.parameter.Format;
 import fr.insee.eno.core.processing.ProcessingStep;
-import fr.insee.eno.core.utils.VtlSyntaxUtils;
+import fr.insee.eno.core.utils.vtl.VtlSyntaxUtils;
 
 import java.util.List;